Move API request func to checkout utils & update ref
